Implements the core data stack in order to store the values persitently

+++ bodymass/Protocols/WeightSelectorDelegate.swift
@@ -9,5 +9,5 @@
 import Foundation
 
 protocol WeightSelectorDelegate: class {
-  func weightChanged(value: Int)
+  func weightChanged(value: Float)
 }
